requirements.txt
streamlit==0.37.0
langchain==0.2.11
langchain-openai==0.1.19
python-dotenv==1.0.1
